// src/types/global.d.ts
export interface AgentVersion {
  id: string;
  hash: string;
  metadata: {
    tools?: string[];
    systemPrompt?: string;
  };
  samplingRate: number;
  createdAt: string;
  updatedAt: string;
}

export interface Agent {
  id: string;
  userId: string;
  projectId: string;
  environmentId: string;
  name: string;
  traceCount: number;
  successPercentage: number | Record<string, number>;
  successRatePeriod?: number;
  lastActive: string;
  createdAt: string;
  updatedAt: string;
  latestVersion?: AgentVersion;
  versions?: AgentVersion[];
}

export interface Pagination {
  total: number;
  count: number;
  page: number;
  totalPages: number;
  hasMore: boolean;
}

export interface AgentsResponse {
  success: boolean;
  agents: Agent[];
  pagination: Pagination;
}

export interface SingleAgentResponse {
  success: boolean;
  id: string;
  userId: string;
  projectId: string;
  environmentId: string;
  name: string;
  traceCount: number;
  successPercentage: Record<string, number>;
  successRatePeriod: number;
  lastActive: string;
  createdAt: string;
  updatedAt: string;
  latestVersion: AgentVersion;
  versions: AgentVersion[];
}